+++ Sources/UberAuth/AuthProviding.swift
@@ -32,6 +32,10 @@ public protocol AuthProviding {
     func execute(authDestination: AuthDestination,
                  prefill: Prefill?,
                  completion: @escaping (Result<Client, UberAuthError>) -> ())
+
+    /// - Throws: `UberAuthError`
+    func execute(authDestination: AuthDestination,
+                 prefill: Prefill?) async throws -> Client
     
     func logout() -> Bool
     
@@ -45,12 +49,14 @@ extension AuthProviding where Self == AuthorizationCodeAuthProvider {
     public static func authorizationCode(presentationAnchor: ASPresentationAnchor = .init(),
                                          scopes: [String] = AuthorizationCodeAuthProvider.defaultScopes,
                                          shouldExchangeAuthCode: Bool = true,
-                                         prompt: Prompt? = nil) -> Self {
+                                         prompt: Prompt? = nil,
+                                         environment: UberEnvironment = .production) -> Self {
         AuthorizationCodeAuthProvider(
             presentationAnchor: presentationAnchor,
             scopes: scopes,
             shouldExchangeAuthCode: shouldExchangeAuthCode,
-            prompt: prompt
+            prompt: prompt,
+            environment: environment
         )
     }
 }