+++ examples/UberSDK/UberSDKTests/UberAuth/AuthorizationCodeResponseParserTests.swift
@@ -105,14 +105,14 @@ final class AuthorizationCodeResponseParserTests: XCTestCase {
     }
 
     func test_parse_invalidUrl_returnsFailure() {
-        
-        let url = URL(filePath: "invalid_scheme://host?code=123")
-        
+
+        let url = URL(string: "invalid-scheme://host")!
+
         let result = responseParser(
             url: url
         )
-        
-        XCTAssertEqual(result, .failure(UberAuthError.invalidResponse))
+
+        XCTAssertEqual(result, .failure(UberAuthError.invalidAuthCode))
     }
     
     func test_parse_authorizationCodeParameter_returnsSuccess() {